+++ .idea/jsLibraryMappings.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="JavaScriptLibraryMappings">
-    <file url="file://$PROJECT_DIR$" libraries="{jquery}" />
+    <file url="file://$PROJECT_DIR$" libraries="{bootstrap, jquery, jquery-3.5.1.slim}" />
   </component>
 </project>
