+++ src/main/webapp/WEB-INF/view/mypage.jsp
@@ -85,6 +85,8 @@
         </c:forEach>
 </div>
 <input type="button" value="예약 취소" onclick="cancelReservation();">
+<input type="button" value="메인 페이지" onclick="location.href='/main'">
+
 <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.1/dist/umd/popper.min.js"
         integrity="sha384-9/reFTGAW83EW2RDu2S0VKaIzap3H66lZH81PoYlFhbGU+6BZp6G7niu735Sk7lN"
         crossorigin="anonymous"></script>
